server/requirements.txt
django==4.0.4
psycopg2==2.9.3
djangorestframework==3.13.1
drf-yasg==1.20.0
django-filter==21.1
coverage==6.3.2
django-cors-headers==3.12.0